models/image_verification.py
import streamlit as st
import cv2
import numpy as np
from PIL import Image
import tempfile
import os
import random

class ImageVerification:
    def __init__(self):
        self.model = None
        self.hazard_classes = {
            'pothole': 0,
            'crack': 1,
            'waterlogging': 2,
            'debris': 3,
            'construction': 4
        }
    
    def load_model(self):
        \"\"\"Load YOLOv8 model - in real implementation, load actual trained model\"\"\"
        try:
            # This would load your trained model in real implementation
            # from ultralytics import YOLO
            # self.model = YOLO(Config.YOLO_MODEL_PATH)
            st.success("âœ… Image verification model loaded (Demo Mode)")
        except ImportError:
            st.warning("ðŸš¨ YOLOv8 not available - using demo verification")
    
    def verify_hazard_image(self, image_file, reported_hazard_type):
        \"\"\"Verify hazard from uploaded image\"\"\"
        # Demo mode - simulate verification
        return self._demo_verification(image_file, reported_hazard_type)
    
    def _demo_verification(self, image_file, reported_hazard_type):
        \"\"\"Demo verification for hackathon\"\"\"
        # Simulate AI processing
        import time
        time.sleep(1)  # Simulate processing time
        
        # Simple image analysis for demo
        try:
            image = Image.open(image_file)
            image.verify()  # Basic image validation
            
            # Mock detection based on file characteristics
            file_size = len(image_file.getvalue())
            confidence = min(0.3 + (file_size / 1000000), 0.85)  # Mock confidence based on file size
            
            # Simulate detection matching (80% chance of matching)
            matches = random.random() > 0.2
            
            return {
                'verified': True,
                'detected_hazards': [{'type': reported_hazard_type, 'confidence': confidence}],
                'matches_reported': matches,
                'overall_confidence': confidence * 100,
                'demo_mode': True
            }
        except Exception as e:
            return {
                'verified': False,
                'detected_hazards': [],
                'matches_reported': False,
                'overall_confidence': 0,
                'error': str(e)
            }
    
    def _get_hazard_name(self, class_id):
        \"\"\"Convert class ID to hazard name\"\"\"
        for name, cid in self.hazard_classes.items():
            if cid == class_id:
                return name
        return 'unknown'
