requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
requests>=2.31.0
plotly>=5.15.0
geopy>=2.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
ultralytics>=8.0.0
scipy>=1.11.0
mapbox>=1.4.0
redis>=5.0.0
networkx>=3.0
sqlite3
